backend/requirements.txt
Django>=4.2,<5.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.4.0